tests: don't unconditionally enable debug mode for tests
This breaks compilation for release-only builds on Mac or Windows.

For platforms where debug vs release affects the library name (e.g.
Mac, Windows), it's invalid to do CONFIG+=debug unless you know that
a debug version of the Qt libraries are available.  Just let the build
system decide whether to build in debug or release mode as normal.

Change-Id: Ibbbdb0fd8395265981f10795b9b539e19ea4fb90
Reviewed-on: http://codereview.qt.nokia.com/4031
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Xizhi Zhu <xizhi.zhu@nokia.com>